#fulltest safe upd bind & udp index header

+++ tests/LinuxFullTest/fulltest_gen_content.py
@@ -1,4 +1,5 @@
 import random, os, shutil
+import fulltest_udp_proto
 
 def get_random_string(length):
     RANDOM_STRING = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
@@ -8,6 +9,9 @@ def get_random_string(length):
     return "".join(arr)
 
 def gen_files(parent_dir, files_count, file_size):
+    if file_size / fulltest_udp_proto.SEND_PACKET_LENGTH >= 256:
+        raise Exception(" gen files is too large! we need remain 1 byte for index of udp")
+
     if os.path.exists(parent_dir) :
         shutil.rmtree(parent_dir)
